Moved migration from userlayer to oskari module since the database table modified is in the core tables
